refactor: move classifySuggestions and filterBatchCoveredSuggestions to suggestion-utils
